backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
cryptography==41.0.7
pydantic-settings==2.1.0
python-dotenv==1.0.0
pypdf==3.17.1
chromadb==0.4.24
openai==1.3.5
SQLAlchemy==2.0.30
